+++ api/src/interfaces/comment.ts
@@ -1,8 +1,0 @@
-import { CreatedAt } from "./blog";
-
-export interface Comment extends CreatedAt {
-  id: string;
-  authorId: string;
-  postId: string;
-  content: string;
-}