+++ api/src/interfaces/payload.d.ts
@@ -1,0 +1,6 @@
+// import { JwtPayload } from "jsonwebtoken";
+import { User } from "./user";
+
+declare module "jsonwebtoken" {
+  export interface JwtPayload extends User {}
+}