+++ api/src/utils/jwt.utils.ts
@@ -1,7 +1,15 @@
-import jwt, { SignOptions, VerifyErrors, JwtPayload } from "jsonwebtoken";
+import jwt, { SignOptions, JwtPayload } from "jsonwebtoken";
 
-export const signJWT = (payload: object, options: SignOptions) => {
-  return jwt.sign(payload, "secretKey", options);
+export const signJWT = async (
+  payload: object,
+  options: SignOptions
+): Promise<string | null> => {
+  return await new Promise((res, rej) => {
+    jwt.sign(payload, "secretKey", options, (err, token) => {
+      if (err || !token) return res(null);
+      return res(token);
+    });
+  });
 };
 
 // jwt SignOptions expiresIn is type StringValue
@@ -9,12 +17,23 @@ export const signJWT = (payload: object, options: SignOptions) => {
   return jwt.sign(payload, "secretKey", { expiresIn });
 }; */
 
-export const verifyJWT = (token: string) => {
-  try {
-    const decoded = jwt.verify(token, "secretKey") as JwtPayload;
-    return { payload: decoded, expired: false };
-  } catch (err) {
-    const e = err as VerifyErrors;
-    return { payload: null, expired: e.name === "TokenExpiredError" };
-  }
+// https://stackoverflow.com/questions/27726066/jwt-refresh-token-flow
+export const verifyJWT = async (
+  token: string
+): Promise<{ payload: null | JwtPayload; expired: boolean }> => {
+  console.log("verifyJWT");
+  console.log("token:", token);
+  return await new Promise((res, rej) => {
+    jwt.verify(token, "secretKey", (err, decoded) => {
+      if (err) {
+        console.log("verifyJWT err:", err);
+        return res({
+          payload: null,
+          expired: !token ? true : err.name === "TokenExpiredError",
+        });
+      } else {
+        return res({ payload: decoded as JwtPayload, expired: false });
+      }
+    });
+  });
 };